Refactor move generation: Introduced MoveGen class for better separation of related functions
